Add support for external types
Allow messages to contain objects that are not generated by protobuf.
In order to do this the user must specifiy a wrapper message type in
the "external" package and provide a serializer and derializer. So
for example if you had a type Bar in namespace Foo you would create
a proto file containing:

package external.Foo;

message Bar {}

The serializer and deserializer must have the following name:

<type>Serializer/<type>Deserializer and must be in the same namespace
as the type being serialized.

They should subclass the ProtoCall::Serialization::Serializer and
ProtoCall::Serialization::Deserializer base classes.