Remove runtime library
Rather than build runtime library, install runtime source
and add it to the common source provided to the user. This
is required as the message.pb.h generate at ProtoCall build
time will be different to the one generated when the user
compiles their .proto files (we are using protobuf extensions).